refactor(skill-progression): extract Layer 3 DB helpers into _db_helpers.py
Move _compute_opponent_factor and _compute_match_performance_modifier
verbatim from skill_progression_service.py (shim) into the sub-package
at app/services/skill_progression/_db_helpers.py.

Both symbols are re-exported from the package __init__.py and imported
back into the shim — callers (Layer 4/5) remain in the shim so
mock.patch LOAD_GLOBAL interception is preserved for all existing tests.

Zero-behaviour-change: 82/82 skill-progression unit tests pass,
7186/7186 full unit suite green.
