+++ app/services/skill_progression/__init__.py
@@ -1,10 +1,11 @@
 """
-skill_progression sub-package — Layers 1 + 2.
+skill_progression sub-package — Layers 1 + 2 + 3.
 
-Layer 1 (_formulas):  pure math, constants
-Layer 2 (_config):    skill key enumeration, baseline lookup, tournament mapping
+Layer 1 (_formulas):    pure math, constants
+Layer 2 (_config):      skill key enumeration, baseline lookup, tournament mapping
+Layer 3 (_db_helpers):  DB-backed opponent factor + match performance modifier
 
-Future layers (DB helpers, EMA engine, views) will be added in subsequent PRs.
+Future layers (EMA engine, views) will be added in subsequent PRs.
 The canonical import surface remains app.services.skill_progression_service (thin shim).
 """
 from ._formulas import (
@@ -20,6 +21,10 @@ from ._config import (
     get_baseline_skills,
     _extract_tournament_skills,
 )
+from ._db_helpers import (
+    _compute_opponent_factor,
+    _compute_match_performance_modifier,
+)
 
 __all__ = [
     "MIN_SKILL_VALUE",
@@ -31,4 +36,6 @@ __all__ = [
     "get_all_skill_keys",
     "get_baseline_skills",
     "_extract_tournament_skills",
+    "_compute_opponent_factor",
+    "_compute_match_performance_modifier",
 ]
